refactoring + external FoliaBridge
