+++ pom.xml
@@ -122,10 +122,17 @@
         <!-- Platform -->
         <!-- Used to support Folia -->
         <dependency>
-            <groupId>io.papermc.paper</groupId>
-            <artifactId>paper-api</artifactId>
-            <version>1.20.2-R0.1-SNAPSHOT</version>
+            <groupId>org.spigotmc</groupId>
+            <artifactId>spigot-api</artifactId>
+            <version>1.20.1-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.promcteam</groupId>
+            <artifactId>ProFoliaBridge</artifactId>
+            <version>1</version>
+            <scope>system</scope>
+            <systemPath>${project.basedir}/lib/ProFoliaBridge.jar</systemPath>
         </dependency>
         <!-- Used for bungee cord -->
         <dependency>
