+++ src/main/java/org/black_ixx/bossshop/core/BSBuy.java
@@ -10,10 +10,10 @@ import org.black_ixx.bossshop.events.BSPlayerPurchaseEvent;
 import org.black_ixx.bossshop.events.BSPlayerPurchasedEvent;
 import org.black_ixx.bossshop.managers.ClassManager;
 import org.black_ixx.bossshop.managers.config.BSConfigShop;
-import org.black_ixx.bossshop.managers.folia.CrossScheduler;
 import org.black_ixx.bossshop.misc.Misc;
 import org.black_ixx.bossshop.misc.ShopItemPurchaseTask;
 import org.black_ixx.bossshop.settings.Settings;
+import org.black_ixx.bossshoppro.folia.CrossScheduler;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
