+++ src/main/java/org/black_ixx/bossshop/listeners/InventoryListener.java
@@ -6,11 +6,11 @@ import org.black_ixx.bossshop.core.BSBuy;
 import org.black_ixx.bossshop.core.BSShop;
 import org.black_ixx.bossshop.core.BSShopHolder;
 import org.black_ixx.bossshop.managers.ClassManager;
-import org.black_ixx.bossshop.managers.folia.CrossScheduler;
 import org.black_ixx.bossshop.misc.MathTools;
 import org.black_ixx.bossshop.misc.Misc;
 import org.black_ixx.bossshop.misc.userinput.BSAnvilHolder;
 import org.black_ixx.bossshop.settings.Settings;
+import org.black_ixx.bossshoppro.folia.CrossScheduler;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event.Result;
 import org.bukkit.event.EventHandler;
