+++ src/main/java/org/black_ixx/bossshop/managers/serverpinging/ServerPingingRunnableHandler.java
@@ -1,8 +1,7 @@
 package org.black_ixx.bossshop.managers.serverpinging;
 
 import org.black_ixx.bossshop.BossShop;
-import org.black_ixx.bossshop.managers.folia.CrossScheduler;
-import org.bukkit.Bukkit;
+import org.black_ixx.bossshoppro.folia.CrossScheduler;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scheduler.BukkitTask;
